// admin_live_mall/src/pages/partner/Mentor/MentorList/data.d.ts
export type TableListItem = {
  key?: number;
  id: string;
  allowanceRatio: string;
  createId: number;
  createTime: string;
  customerNum: number;
  image: string;
  isDefault: boolean;
  modifyId: string;
  modifyTime: string;
  name: string;
  phone: string;
  remarkName: string;
  settledCommission: string;
  teamNum: number;
  toBeSettledCommission: number | string;
};